add node-sass builder + update dependencies + doc
