+++ .gsk/conf/.sass-lint.yml
@@ -1,0 +1,104 @@
+# sass-lint config generated by make-sass-lint-config v0.1.2
+#
+# The following scss-lint Linters are not yet supported by sass-lint:
+# Compass::PropertyWithMixin, ElsePlacement, PropertyCount, SelectorDepth
+# UnnecessaryParentReference, VendorPrefixes
+#
+# The following settings/values are unsupported by sass-lint:
+# Linter Indentation, option "allow_non_nested_indentation"
+# Linter Indentation, option "character"
+# Linter SpaceAfterPropertyColon, option "style" with value "at_least_one_space"
+# Linter SpaceBeforeBrace, option "allow_single_line_padding"
+
+files:
+  include: '**/*.scss'
+options:
+  formatter: stylish
+  merge-default-rules: false
+rules:
+  border-zero: 0
+  brace-style:
+    - 1
+    - allow-single-line: true
+  class-name-format: 0
+  clean-import-paths:
+    - 1
+    - filename-extension: false
+      leading-underscore: false
+  empty-line-between-blocks:
+    - 1
+    - ignore-single-line-rulesets: true
+  extends-before-declarations: 1
+  extends-before-mixins: 1
+  final-newline:
+    - 1
+    - include: true
+  force-attribute-nesting: 0
+  force-element-nesting: 0
+  force-pseudo-nesting: 0
+  function-name-format:
+    - 1
+    - allow-leading-underscore: false
+      convention: hyphenatedlowercase
+  hex-length: 0
+  hex-notation: 0
+  id-name-format: 0
+  indentation:
+    - 1
+    - size: 2
+  leading-zero: 0
+  mixin-name-format:
+    - 1
+    - allow-leading-underscore: false
+      convention: hyphenatedlowercase
+  mixins-before-declarations: 1
+  nesting-depth:
+    - 1
+    - max-depth: 5
+  no-color-keywords: 1
+  no-color-literals: 1
+  no-css-comments: 1
+  no-debug: 1
+  no-duplicate-properties: 0
+  no-empty-rulesets: 1
+  no-ids: 0
+  no-important: 0
+  no-invalid-hex: 1
+  no-mergeable-selectors: 1
+  no-misspelled-properties: 0
+  no-qualifying-elements:
+    - 1
+    - allow-element-with-attribute: true
+      allow-element-with-class: false
+      allow-element-with-id: false
+  no-trailing-zero: 1
+  no-url-protocols: 0
+  placeholder-in-extend: 0
+  placeholder-name-format: 0
+  property-sort-order: 0
+  quotes: 0
+  shorthand-values: 1
+  single-line-per-selector: 1
+  space-after-bang:
+    - 1
+    - include: false
+  space-after-colon: 1
+  space-after-comma: 1
+  space-before-bang:
+    - 1
+    - include: true
+  space-before-brace:
+    - 1
+    - include: true
+  space-before-colon: 1
+  space-between-parens:
+    - 1
+    - include: false
+  trailing-semicolon: 1
+  url-quotes: 1
+  variable-for-property: 0
+  variable-name-format:
+    - 1
+    - allow-leading-underscore: false
+      convention: hyphenatedlowercase
+  zero-unit: 1
