+++ .gsk/conf/readme.md
@@ -11,9 +11,12 @@ de votre projet. Cependant, si vous modifiez systématiquement les
 configurations par défaut, n'hésitez pas à proposer une merge request au
 starter kit pour que l'on en discute.
 
+* `.editorconfig`  Configuration commune pour le [format des fichier pour les IDE](http://editorconfig.org/)
 * `.jsbeautifyrc`  Configuration pour le [formatage automatique du code HTML](https://github.com/beautify-web/js-beautify)
 * `.eslint.json`   Configuration pour [le linting et la vérification de syntaxe des fichiers JS](http://eslint.org)
 * `.lesshintrc`    Configuration pour le [linting des fichiers LESS](https://github.com/lesshint/lesshint)
-* `.scss-lint.yml` Configuration pour le [linting des fichiers SCSS](https://github.com/brigade/scss-lint/)
+* `.scss-lint.yml` Configuration pour le [linting des fichiers SCSS (pour compass)](https://github.com/brigade/scss-lint/)
+* `.sass-lint.yml` Configuration pour le [linting des fichiers SCSS (pour eyeglass)](https://github.com/sasstools/sass-lint)
 * `.stylintrc`     Configuration pour le [linting des fichiers Stylus](https://rosspatton.github.io/stylint/)
 * `config.rb`      Configuration pour la [compilation Compass](http://compass-style.org/)
+* `Gemfile` [Config pour dependances ruby (si on utilise Compass)](http://bundler.io/gemfile.html)
