--- Gemfile
@@ -1,6 +1,0 @@
-source 'https://rubygems.org'
-
-gem 'compass',   '1.0.3'  # Compass
-gem 'sass',      '3.4.18' # Sass, évite les mauvaises surprises
-gem 'oily_png',  '1.2.0'  # OilyPNG, accélère la création des sprites
-gem 'scss_lint', '0.42.2' # Linting des fichiers de style
