+++ package.json
@@ -3,11 +3,7 @@
   "homepage": "https://github.com/cleverage/garden-starter-kit",
   "title": "Garden Starter Kit",
   "description": "Set de base pour les projets Garden statiques.",
-<<<<<<< d464be779a04eb7d45f44b437c2a220e248805df
-  "version": "2.0.0-beta7",
-=======
-  "version": "2.0.0-beta6",
->>>>>>> Replace jshint and jscs by eslint #81
+  "version": "2.0.0-beta8",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -32,6 +28,7 @@
     "gulp-less": "3.3.0",
     "gulp-plumber": "1.1.0",
     "gulp-postcss": "6.4.0",
+    "gulp-sass": "3.1.0",
     "gulp-sourcemaps": "2.4.1",
     "gulp-stylus": "2.6.0",
     "gulp-svgstore": "6.1.0",
@@ -39,6 +36,8 @@
     "gulp-util": "3.0.8",
     "jquery": "3.2.1",
     "lazypipe": "1.0.1",
+    "node-sass": "4.5.0",
+    "node-sass-import": "1.1.1",
     "merge-stream": "1.0.1",
     "require-dir": "0.3.1",
     "run-sequence": "1.2.2",
@@ -48,7 +47,7 @@
     "vinyl-source-stream": "1.1.0",
     "webpack": "2.3.1",
     "webpack-stream": "3.2.0",
-    "yargs": "6.6.0"
+    "yargs": "7.0.2"
   },
   "devDependencies": {
     "browser-sync": "2.18.8",
@@ -60,20 +59,20 @@
     "gulp-hb": "6.0.2",
     "gulp-hbs": "0.2.0",
     "gulp-help": "1.6.1",
-    "gulp-lesshint": "2.1.0",
+    "gulp-lesshint": "3.0.0",
     "gulp-markdown": "1.2.0",
     "gulp-newer": "1.3.0",
     "gulp-prompt": "0.2.0",
     "gulp-rename": "1.2.2",
+    "gulp-sass-lint": "1.3.2",
     "gulp-scss-lint": "0.4.0",
     "gulp-sftp": "0.1.5",
     "gulp-stylint": "4.0.0",
+    "gulp-svgstore": "6.0.0",
     "gulp-twig": "0.7.0",
     "kss": "3.0.0-beta.18",
-    "nproxy": "stephane-tessier/nproxy#1.7.1"
-  },
-  "scripts": {
-    "install": "bundle install --path .gems"
+    "nproxy": "stephane-tessier/nproxy#1.7.1",
+    "sass-lint": "1.10.2"
   },
   "engines": {
     "node": ">= 5.0.3",
