+++ src/css/doc.scss
@@ -17,7 +17,6 @@ $color-fg-dark-cell: #FFFFFF;
 $color-fg-light-cell: $darker-grey;
 $color-bg-light-cell: #FFFFFF;
 
-
 // Grid
 $column: 4rem;
 $gap: 1rem;
