+++ src/css/styleguide.md
@@ -1,4 +1,4 @@
 Guide des styles
 ===============================================================================
 
-Page d'accueil du guide des styles en markdown.
+Page d’accueil du guide des styles en markdown.
