+++ src/css/styles.scss
@@ -3,61 +3,60 @@
 // --------------------------
 
 /// Générez un guide des styles
- //
- // Si vous commentez votre code en utilisant la syntaxe
- // [KSS](https://github.com/kss-node/kss/blob/spec/SPEC.md),
- // le Starter Kit generera un guide des styles dans la
- // documentation a chaque build du projet.
- //
- // Styleguide Introduction
-
+//
+// Si vous commentez votre code en utilisant la syntaxe
+// [KSS](https://github.com/kss-node/kss/blob/spec/SPEC.md),
+// le Starter Kit generera un guide des styles dans la
+// documentation a chaque build du projet.
+//
+// Styleguide Introduction
 
 /// Une palette de couleur
- //
- // Si vous voulez documenter des couleurs, vous pouvez
- // le faire de cette manière grace au block Colors.
- //
- // Colors:
- // denim          : #1e5bc6
- // flamingo       : #f15145
- // alert          : #f33121 - red for error
- // $color-black   : #000    - real black
- //
- // Styleguide Introduction.Couleur
+//
+// Si vous voulez documenter des couleurs, vous pouvez
+// le faire de cette manière grace au block Colors.
+//
+// Colors:
+// denim          : #1e5bc6
+// flamingo       : #f15145
+// alert          : #f33121 - red for error
+// $color-black   : #000    - real black
+//
+// Styleguide Introduction.Couleur
 
 /// Font d’icon Fontello
- //
- // De la même façon que pour les fonts grace au block
- // Icons vous pouvez documenter les icons de la font.
- //
- // Implique que la font d'icon se nomme fontello
- // Sinon il faut changer celle-ci dans doc.scss
- //
- // Icons:
- // left    : ←
- // up      : ↑
- // right   : →
- // down    : ↓
- // comment : 💬 - bulle
- //
- // Styleguide Introduction.Icons
-
+//
+// De la même façon que pour les fonts grâce au block
+// Icons vous pouvez documenter les icons de la font.
+//
+// Implique que la font d’icon se nomme fontello
+// Sinon il faut changer celle-ci dans doc.scss
+//
+// Icons:
+// left    : ←
+// up      : ↑
+// right   : →
+// down    : ↓
+// comment : 💬 - bulle
+//
+// Styleguide Introduction.Icons
 
 /// Exemple UI
- //
- // Markup: ../html/partials/_exemple.twig
- //
- // Styleguide Introduction.UI
+//
+// Markup: ../html/partials/_exemple.twig
+//
+// Styleguide Introduction.UI
 
 /// SVG
- //
- // Markup:
- // <svg style="height:120px;width:120px;fill:#000">
- //   <use xlink:href="#symbol-fish"/>
- // </svg>
- //
- // Styleguide Introduction.Svg
+//
+// Markup:
+// <svg style="height:120px;width:120px;fill:#000">
+//   <use xlink:href="#symbol-fish"/>
+// </svg>
+//
+// Styleguide Introduction.Svg
 
 body {
   font: 1em sans-serif;
+  background: url('../img/clever-age-logo.png') no-repeat top right;
 }
